+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to `when` are documented here.
 
 - Added support for multiple target locations.
 - Added timezone name to short output.
+- Added `--json` flag to enable JSON output.
 
 ## 0.2.0